Added ability to save detected roadSigns and added some changes for models
